taskController: Updated controller functions.

+++ server/controller/taskController.js
@@ -7,9 +7,6 @@ export const getTasks = async (req, res) => {
       ...task,
       id: task._id.toString(),
     }));
-    if (formattedTasks.length === 0) {
-      return res.status(404).json({ error: "No tasks found" });
-    }
 
     res.status(200).json({
       message: "Tasks fetched successfully",
@@ -24,7 +21,7 @@ export const getTasks = async (req, res) => {
 export const postTasks = async (req, res) => {
   try {
     const { title } = req.body;
-    if (!title) return res.status(400).send({ error: "Title is required" });
+    if (!title) return res.status(400).json({ error: "Title is required" });
 
     const newTask = new Task({
       title,
@@ -34,7 +31,10 @@ export const postTasks = async (req, res) => {
     });
 
     await newTask.save();
-    res.status(201).send(newTask);
+    res.status(201).json({
+      message: "Tasks added successfully",
+      task: newTask,
+    });
   } catch (error) {
     console.error("Error in postTasks:", error);
     res.status(500).json({ message: "Server error" });
@@ -45,16 +45,23 @@ export const putTasks = async (req, res) => {
   try {
     const { id } = req.params;
     const { title, completed } = req.body;
+    if (!id || !title || completed === undefined) {
+      return res
+        .status(400)
+        .json({ error: "ID, title, and completed status are required" });
+    }
 
     const updatedTask = await Task.findByIdAndUpdate(
       id,
       { title, completed, updatedAt: new Date() },
       { new: true }
     );
-
     if (!updatedTask) return res.status(404).send({ error: "Task not found" });
 
-    res.status(200).send(updatedTask);
+    res.status(200).json({
+      message: "Tasks updated successfully",
+      task: updatedTask,
+    });
   } catch (error) {
     console.error("Error in putTasks:", error);
     res.status(500).json({ message: "Server error" });
@@ -64,12 +71,12 @@ export const putTasks = async (req, res) => {
 export const deleteTasks = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!id) return res.status(400).json({ error: "ID required" });
 
     const deletedTask = await Task.findByIdAndDelete(id);
-
     if (!deletedTask) return res.status(404).send({ error: "Task not found" });
 
-    res.status(200).send({ message: "Task deleted successfully" });
+    res.status(200).json({ message: "Task deleted successfully" });
   } catch (error) {
     console.error("Error in deleteTasks:", error);
     res.status(500).json({ message: "Server error" });
